+++ src/renderer/src/context/Projects/api/createCommunityVideosVAMEProject.ts
@@ -7,7 +7,7 @@ type CreateCommunityVideosProps = {
 }
 
 export const createCommunityVideosVAMEProject = async (data: CreateCommunityVideosProps) => {
-  const result = await post<ProjectType>('community_videos', { ...data })
+  const result = await post<ProjectType>('community-videos', { ...data })
 
   if (result.success) {
     return result.data
